+++ example/lib/constants.dart
@@ -2,7 +2,7 @@ class Constants {
   static const String bugBuckBunnyVideoUrl =
       'https://cdn.jsdelivr.net/npm/big-buck-bunny-1080p@0.0.6/video.mp4';
   static const String forBiggerBlazesUrl =
-      'https://commondatastorage.googleapis.com/gtv-videos-bucket/sample/ForBiggerBlazes.mp4';
+      'https://cdn.jsdelivr.net/gh/akash-paliwal/ExoPlayer-Demo/ForBiggerBlazes.mp4';
   static const String fileTestVideoUrl = 'testvideo.mp4';
   static const String fileTestVideoEncryptUrl = 'testvideo_encrypt.mp4';
   static const String networkTestVideoEncryptUrl =
@@ -27,9 +27,9 @@ class Constants {
   static const String elephantDreamVideoUrl =
       'https://archive.org/download/ElephantsDream/ed_1024_512kb.mp4';
   static const String forBiggerJoyridesVideoUrl =
-      'https://commondatastorage.googleapis.com/gtv-videos-bucket/sample/ForBiggerJoyrides.mp4';
+      'https://cdn.jsdelivr.net/gh/lucpotage/video-demo@master/video/ForBiggerJoyrides.mp4';
   static const String verticalVideoUrl =
-      'https://commondatastorage.googleapis.com/gtv-videos-bucket/sample/ForBiggerJoyrides.mp4';
+      'https://cdn.jsdelivr.net/gh/lucpotage/video-demo@master/video/ForBiggerJoyrides.mp4';
   static String logo = 'logo.png';
   static String placeholderUrl = 'https://picsum.photos/id/10/1000/1000';
   static String elephantDreamStreamUrl =